fix(hipdnn): address review of the descriptor-backed engine work
Correctness, conformance and robustness fixes from review of ALMIOPEN-2401.

The conv matcher never related x, w and y dimensions. It validated each
operand on its own, then bound and returned true, so a shape-inconsistent
graph was accepted and the kernel read and wrote out of bounds: a filter with
fewer channels than x read past the end of w (prepare() takes c from x but
indexes w with it, which is also why grouped conv was silently claimed and
computed wrong), a y smaller than {N,K,H-R+1,W-S+1} was written past its end
because the grid is sized from x and w alone, and r > h wrapped p negative
into an unsigned grid size. Nothing upstream covers this: the backend's
validateConvolutionFinalize only asserts the attributes are set, and the
frontend's dim checks are optional to reach. Three comparisons close it, and
the two regression tests fail against the previous matcher.

Conv index arithmetic was 32-bit and overflowed on shapes the matcher admits
(N*C*H*W passes 2^31 at 32x256x256x256), which is UB and in practice a
negative index. Widened to int64_t in both the kernel and the grid-size math,
so the two cannot disagree.

Restores RFC conformance in two places the format had drifted. `schema` is
required again on all six types, because RFC 0020 4.2 lists it as required
and a spec-conformant file was failing with `unknown key 'schema'`; the
filename suffix still types the file before it is opened and `schema` is now
cross-checked against it, so a .kmd.json renamed .uhd.json is caught by name
instead of surfacing as `unknown key 'fields'`. Five KDP keys move to the
spellings both RFCs use: matchers, engine, dispatch, kernelDescriptors,
kernel_source. Neither old spelling is accepted; aliasing would be the second
source of truth this work exists to avoid.

The supported version moves into the FileType row, one per type. A single
shared pair undercut the reason for requiring `version` everywhere, since the
first type to reach 1.1 would have made the gate accept 1.1 for all six.

`arch` was the only string in the format with no validation, and the one
whose typo is invisible: "GFX942" parsed cleanly and then declined on every
device, logging the same INFO line a healthy cross-arch install logs. Now
rejects empty, malformed and duplicate entries.

The descriptor directory resolves relative to the loaded module before
falling back to the configure-time prefix, so a DESTDIR-staged, relocated or
repackaged install finds its own descriptors instead of advertising zero
engines. Verified by moving an install tree and confirming it still loads.

Also: the walk survives an unreadable subdirectory instead of truncating,
warns on .JSON and .jsonc instead of dropping them silently, and cannot throw
on a Windows filename; the orphan-pack diagnostic sorts; operation_is_sub
gets a fresh id rather than recycling the retired sub graph-matcher's; and
the docs that claimed arch and sdk_version were unparsed, or that a load-time
arch gate ships, now match the code.

Tests: the DIV decline case is device-guarded like its five siblings, two
tautological pack assertions now compare against the set's engine id, the
shipped ConvFwd set is loaded and pinned by a unit test for the first time,
subtraction direction is asserted without a GPU, and temp dirs use mkdtemp.

SDK 308, provider 512 + 1 pre-existing skip, GPU 14/14 on gfx90a;
flag-OFF 106 + 428 with no arch_content staged.

+++ dnn-providers/hip-kernel-provider/src/engines/kernel_ingestor_engine/kernels/ConvFwd.cpp
@@ -16,16 +16,21 @@
 // HIP_PLUGIN_CONV_TYPE: a _Float16 accumulator loses too much precision for a reference
 // a CPU float reference is compared against.
 
+#include <cstdint>
+
 extern "C" __global__ void ConvFwd(const HIP_PLUGIN_CONV_TYPE* x,
                                    const HIP_PLUGIN_CONV_TYPE* w,
                                    HIP_PLUGIN_CONV_TYPE* y,
                                    int n, int c, int h, int width, int k, int r, int s)
 {
-    const int p = h - r + 1;
-    const int q = width - s + 1;
-    const int total = n * k * p * q;
+    // int64_t: n*k*p*q can exceed 2^31 for shapes the matcher admits (element count is
+    // unbounded); 32-bit arithmetic here wrapped, defeating the `index >= total` guard
+    // and turning it into an out-of-bounds read/write.
+    const int64_t p = h - r + 1;
+    const int64_t q = width - s + 1;
+    const int64_t total = static_cast<int64_t>(n) * k * p * q;
 
-    const int index = static_cast<int>(blockIdx.x * blockDim.x + threadIdx.x);
+    const int64_t index = static_cast<int64_t>(blockIdx.x) * blockDim.x + threadIdx.x;
     if(index >= total)
     {
         return;
@@ -33,14 +38,14 @@ extern "C" __global__ void ConvFwd(const HIP_PLUGIN_CONV_TYPE* x,
 
     // Unravel the linear index over (n, k, p, q), q fastest -- the same order y is
     // stored in, so index also addresses y directly.
-    int remaining = index;
-    const int qOut = remaining % q;
+    int64_t remaining = index;
+    const int64_t qOut = remaining % q;
     remaining /= q;
-    const int pOut = remaining % p;
+    const int64_t pOut = remaining % p;
     remaining /= p;
-    const int kOut = remaining % k;
+    const int64_t kOut = remaining % k;
     remaining /= k;
-    const int nOut = remaining;
+    const int64_t nOut = remaining;
 
     float accumulator = 0.0f;
     for(int cIn = 0; cIn < c; ++cIn)
@@ -49,10 +54,10 @@ extern "C" __global__ void ConvFwd(const HIP_PLUGIN_CONV_TYPE* x,
         {
             for(int sIn = 0; sIn < s; ++sIn)
             {
-                const int hIn = pOut + rIn;
-                const int wIn = qOut + sIn;
-                const int xIndex = ((nOut * c + cIn) * h + hIn) * width + wIn;
-                const int wIndex = ((kOut * c + cIn) * r + rIn) * s + sIn;
+                const int64_t hIn = pOut + rIn;
+                const int64_t wIn = qOut + sIn;
+                const int64_t xIndex = ((nOut * c + cIn) * h + hIn) * width + wIn;
+                const int64_t wIndex = ((kOut * c + cIn) * r + rIn) * s + sIn;
                 accumulator += static_cast<float>(x[xIndex]) * static_cast<float>(w[wIndex]);
             }
         }
